+++ src/cmd_spanr/mod.rs
@@ -3,6 +3,7 @@
 pub mod combine;
 pub mod compare;
 pub mod convert;
+pub mod count;
 pub mod cover;
 pub mod coverage;
 pub mod genome;