+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased - ReleaseDate
 
+* Add `spanr count`
 * Add `--detailed` to `spanr coverage`
+* Use `Box<dyn std::error::Error>`
 
 ## 0.6.5 - 2022-04-22